+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 0.0.6
+ - Fixed problem with non quadratic image sizes.
+ - Not saying when it is using the BigTIFF format any more.
+
 # 0.0.5
  - added approximate conversion from emission wavelength to color.